:sparkles: Rightbarのオンラインの友達 &  Sidebarの友達 のダミーデータ更新 #19
