+++ package.json
@@ -20,7 +20,8 @@
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.5",
     "typescript": "^4.6.4",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "webpack": "^5.72.1"
   },
   "scripts": {
     "start": "react-scripts start",
